Implemented logic to retrieve issues at the right time
Implemented a retry callback for superagent. This takes in a condition that is evaluated on every retry - Retry until the status of the scan of the current session id is 'success' or 'failed'.

We no longer need to sleep for a fixed duration before fetching issues. The request to sonarqube web api handles this.

Issue #8 fixed